fix: address review findings
